backend/requirements.txt
# ── Django y DRF ───────────────────────────────────────────
Django==5.0.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.1

# ── Base de Datos ───────────────────────────────────────────
psycopg2-binary==2.9.9

# ── Caché ───────────────────────────────────────────────────
django-redis==5.4.0
redis==5.0.3

# ── Almacenamiento ─────────────────────────────────────────
django-storages==1.14.2
boto3==1.34.0           # Opcional: para S3
Pillow==10.3.0          # Procesamiento de imágenes

# ── Servidor ───────────────────────────────────────────────
gunicorn==22.0.0
whitenoise==6.6.0

# ── Pagos ───────────────────────────────────────────────────
stripe==9.5.0

# ── Utilidades ─────────────────────────────────────────────
python-decouple==3.8     # Variables de entorno alternativo
python-dotenv==1.0.1     # Carga automática de .env en desarrollo

# ── Seguridad y 2FA ────────────────────────────────────
django-ratelimit==4.1.0  # Rate limiting adicional por vista
argon2-cffi>=23.1.0      # Hashing Argon2 para contraseñas (más seguro que PBKDF2)
django-ipware>=7.0.0     # Extracción segura de IP real del cliente (anti-spoofing)
django-otp==1.4.0        # 2FA TOTP para el panel de administración
qrcode[pil]==7.4.2       # Genera QR para configuración TOTP (Google Authenticator)

# ── Interfaz administrativa ─────────────────────────────
django-jazzmin==2.6.0    # Tema moderno para el panel de administración Django

# ── Cola asíncrona (Celery + Redis broker) ───────────────
celery==5.4.0
django-celery-beat==2.7.0        # Tareas periódicas almacenadas en DB
django-celery-results==2.5.1     # Resultados de tareas en DB
requests==2.32.3                 # HTTP client para llamadas a APIs de proveedores
certifi>=2024.2.2                # Bundle de CAs — necesario en imágenes slim sin CA store del SO

# ── Cifrado de credenciales de proveedores ───────────────
cryptography==43.0.0             # Fernet: cifrado simétrico AES-128-CBC
